Disable page scroll when modal is open
