--- src/components/backdrop/index.ts
@@ -1,1 +1,0 @@
-export * from './backdrop';
