+++ src/components/index.ts
@@ -1,6 +1,5 @@
 export * from './alerts';
 export * from './avatar';
-export * from './backdrop';
 export * from './breadcrumb';
 export * from './button';
 export * from './cards';
