+++ src/components/modal/modal.stories.tsx
@@ -1,28 +1,45 @@
-import { type Meta, type Story } from '@storybook/react';
-import React, { type ReactNode } from 'react';
+import { type Meta, type StoryObj } from '@storybook/react';
+import React, { useState } from 'react';
+import { ButtonText } from '../button';
 import { Modal, type ModalProps } from './modal';
 
-export default {
+const meta: Meta<typeof Modal> = {
     title: 'Components/Modal',
     component: Modal,
-} as Meta;
-
-const Template: Story<ModalProps> = (args) => <Modal {...args} />;
-
-const TestContent: ReactNode = (
-    <div>
-        Lorem ipsum dolor sit amet consectetur adipisicing elit. Maxime mollitia, molestiae quas vel sint commodi
-        repudiandae consequuntur voluptatum laborum numquam blanditiis harum quisquam eius sed odit fugiat iusto fuga
-        praesentium optio, eaque rerum! Provident similique accusantium nemo autem. Veritatis obcaecati tenetur iure
-        eius earum ut molestias architecto voluptate aliquam nihil, eveniet aliquid culpa officia aut! Impedit sit sunt
-        quaerat, odit, tenetur error, harum nesciunt ipsum debitis quas aliquid. Reprehenderit, quia. Quo neque error
-        repudiandae fuga?
-    </div>
-);
-
-export const Default = Template.bind({});
-Default.args = {
-    children: TestContent,
-    title: 'Test Modal',
-    subtitle: 'Subtitle',
 };
+
+type Story = StoryObj<typeof Modal>;
+
+const ModalDefault = ({ onClose, onInteractOutside, ...props }: ModalProps) => {
+    const [isOpen, setIsOpen] = useState(false);
+
+    return (
+        <>
+            <ButtonText label="Open dialog" onClick={() => setIsOpen(true)} />
+            <Modal isOpen={isOpen} onClose={() => setIsOpen(false)} {...props}>
+                <div style={{ padding: 16, display: 'flex', flexDirection: 'column', gap: 32 }}>
+                    <p>
+                        Lorem ipsum dolor sit amet consectetur adipisicing elit. Maxime mollitia, molestiae quas vel
+                        sint commodi repudiandae consequuntur voluptatum laborum numquam blanditiis harum quisquam eius
+                        sed odit fugiat iusto fuga praesentium optio, eaque rerum!
+                    </p>
+                    <div style={{ height: 300, width: '100%', backgroundColor: 'lightblue' }} />
+                    <p>
+                        Provident similique accusantium nemo autem. Veritatis obcaecati tenetur iure eius earum ut
+                        molestias architecto voluptate aliquam nihil, eveniet aliquid culpa officia aut! Impedit sit
+                        sunt quaerat, odit, tenetur error, harum nesciunt ipsum debitis quas aliquid. Reprehenderit,
+                        quia.
+                    </p>
+                    <div style={{ height: 300, width: '100%', backgroundColor: 'lightblue' }} />
+                    <p>Quo neque error repudiandae fuga?</p>
+                </div>
+            </Modal>
+        </>
+    );
+};
+
+export const Default: Story = {
+    render: (args) => <ModalDefault {...args} />,
+};
+
+export default meta;
